+++ BruteForce/Sources/App/SceneDelegate.swift
@@ -1,0 +1,32 @@
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    
+    var window: UIWindow?
+    
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(frame: windowScene.coordinateSpace.bounds)
+        window?.windowScene = windowScene
+        window?.rootViewController = ViewController()
+        window?.makeKeyAndVisible()
+    }
+    
+    func sceneDidDisconnect(_ scene: UIScene) {
+    }
+    
+    func sceneDidBecomeActive(_ scene: UIScene) {
+    }
+    
+    func sceneWillResignActive(_ scene: UIScene) {
+    }
+    
+    func sceneWillEnterForeground(_ scene: UIScene) {
+    }
+    
+    func sceneDidEnterBackground(_ scene: UIScene) {
+    }
+}
+
