+++ BruteForce/Sources/Scenes/Controller/ViewController.swift
@@ -1,0 +1,33 @@
+import UIKit
+
+class ViewController: UIViewController {
+    
+    
+    
+    lazy var button: UIButton = {
+        let button = UIButton()
+        button.setTitle("Button", for: .normal)
+        button.backgroundColor = .systemRed
+        button.layer.cornerRadius = 15
+        button.translatesAutoresizingMaskIntoConstraints = false
+        return button
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.addSubview(button)
+        view.backgroundColor = .white
+        
+        configure()
+    }
+
+    func configure() {
+        
+        NSLayoutConstraint.activate([
+            button.centerYAnchor.constraint(equalTo: view.centerYAnchor),
+            button.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            button.widthAnchor.constraint(equalToConstant: 100)
+        ])
+        
+    }
+}
