+++ BruteForce/Sources/Scenes/View/ViewProperties.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class ViewProperties: UIViewController {
+
+}
