--- BruteForce/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  BruteForce
-//
-//  Created by Mark Wilhelm on 11.07.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
